microservice

+++ customer/customer/src/main/java/com/customer/customer/repository/CustomerRepository.java
@@ -4,9 +4,12 @@ import com.customer.customer.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer,Long> {
     Customer findByEmail(String email);
 
     Customer findByMobileNumber(String mobileNumber);
+
 }
